+++ src/Pyro4/util.py
@@ -309,7 +309,7 @@ class SerializerBase(object):
             return converter(classname, data)
         if "__" in classname:
             raise errors.SecurityError("refused to deserialize types with double underscores in their name: " + classname)
-        # because of efficiency reasons the constructors below are hardcoded here instead of added on a per-class basis to the dict-to-class registry
+        # for performance, the constructors below are hardcoded here instead of added on a per-class basis to the dict-to-class registry
         if classname.startswith("Pyro4.core."):
             from Pyro4 import core  # XXX circular
             if classname == "Pyro4.core.URI":
@@ -665,6 +665,30 @@ def excepthook(ex_type, ex_value, ex_tb):
     """An exception hook you can use for ``sys.excepthook``, to automatically print remote Pyro tracebacks"""
     traceback = "".join(getPyroTraceback(ex_type, ex_value, ex_tb))
     sys.stderr.write(traceback)
+
+
+def fixIronPythonExceptionForPickle(exceptionObject, addAttributes):
+    """
+    Function to hack around a bug in IronPython where it doesn't pickle
+    exception attributes. We piggyback them into the exception's args.
+    Bug report is at https://github.com/IronLanguages/main/issues/943
+    Bug is still present in Ironpython 2.7.7
+    """
+    if hasattr(exceptionObject, "args"):
+        if addAttributes:
+            # piggyback the attributes on the exception args instead.
+            ironpythonArgs = vars(exceptionObject)
+            ironpythonArgs["__ironpythonargs__"] = True
+            exceptionObject.args += (ironpythonArgs,)
+        else:
+            # check if there is a piggybacked object in the args
+            # if there is, extract the exception attributes from it.
+            if len(exceptionObject.args) > 0:
+                piggyback = exceptionObject.args[-1]
+                if type(piggyback) is dict and piggyback.get("__ironpythonargs__"):
+                    del piggyback["__ironpythonargs__"]
+                    exceptionObject.args = exceptionObject.args[:-1]
+                    exceptionObject.__dict__.update(piggyback)
 
 
 __exposed_member_cache = {}