+++ src/Pyro4/util.py
@@ -635,7 +635,7 @@ try:
         ver = serpent.__version__
     ver = tuple(map(int, ver.split(".")))
     if ver < (1, 16):
-        raise RuntimeError("requires serpent 1.16 or better")
+        raise RuntimeError("requires serpent 1.17 or better")
     _ser = SerpentSerializer()
     _serializers["serpent"] = _ser
     _serializers_by_id[_ser.serializer_id] = _ser